+++ src/component/location/Location.java
@@ -12,10 +12,10 @@ public class Location {
 	private MainCharacter owner;
 //	private ArrayList<E> *TODO:for group minion that is on the tile*
 	private ArrayList<Minion> minionOnLocation;
-	private int incomePerRound;
-	private int cost; 
+	private double incomePerRound;
+	private double cost; 
 	private boolean isPossessable;
-	public Location(String name, String description, int income, int cost) {
+	public Location(String name, String description, double income, double cost) {
 		this.name = name;
 		this.description = description;
 		this.incomePerRound = income*MainCharacter.M;
@@ -79,13 +79,13 @@ public class Location {
 		}
 		return false;
 	}
-	public int getIncomePerRound() {
+	public double getIncomePerRound() {
 		return incomePerRound;
 	}
 	public void setIncomePerRound(int incomePerRound) {
 		this.incomePerRound = incomePerRound;
 	}
-	public int getCost() {
+	public double getCost() {
 		return cost;
 	}
 	public void setCost(int cost) {
